+++ src/main/java/com/swef/cookcode/recipe/service/RecipeService.java
@@ -35,6 +35,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -108,6 +109,7 @@ public class RecipeService {
                 .build();
     }
 
+    @Async
     public void deleteCancelledFiles(RecipeCreateRequest request) {
         util.deleteFilesInS3(request.getDeletedThumbnails());
         request.getSteps().forEach(step -> {
